VIOLET-KICK-CHAT/requirements.txt
pygame>=2.5.0
Pillow>=10.0.0
requests>=2.30.0
gtts>=2.3.0
uvicorn==0.22.0
fastapi==0.95.2
httpx==0.24.1
python-dotenv==1.0.0
python-socketio==5.9.4
websockets==11.0.3
sqlalchemy==2.1.3
aiosqlite==0.18.0